requirements.txt
# Basic requirements
numpy>=1.20.0
matplotlib>=3.4.0
pytest>=6.2.5

# Voice recognition
SpeechRecognition>=3.8.1
pyaudio>=0.2.11

# AI/NLP processing
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=0.19.0

# OpenSCAD interface
solidpython>=1.1.0

# Database
sqlalchemy>=1.4.0
alembic>=1.7.0

# UI and utilities
tqdm>=4.62.0
colorama>=0.4.4
rich>=10.0.0
click>=8.0.0

# Additional dependencies
pydantic>=1.8.2
pillow>=8.3.1